% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{sparselca}
\alias{sparselca-package}
\alias{sparselca}
\title{Bayesian LCA with Sparse Finite Mixtures}
\description{
Uses MCMC to fit latent class models to binary data.
}
\author{
Deborah Kunkel \email{dekunke@clemson.edu}
}
